requirements.txt
keyring
cryptography
google-generativeai
rich
pyperclip
keyboard
